src/test/java/pl/burbaniak/classicdatabase/api/CustomerControllerTest.java
package pl.burbaniak.classicdatabase.api;

import org.junit.jupiter.api.Test;
import org.junit.runner.RunWith;
import org.mockito.Mockito;
import org.springframework.beans.factory.annotation.Autowired;

import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.RequestBuilder;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
import pl.burbaniak.classicdatabase.manager.CustomerManager;

import static org.junit.jupiter.api.Assertions.*;

@RunWith(SpringRunner.class)
@WebMvcTest(value = CustomerController.class)
class CustomerControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private CustomerManager customerManager;

    @Test
    void shouldReturnCustomerWithGivenId() throws Exception {

        mockMvc.perform(MockMvcRequestBuilders.get("/customer-srv/customers/114"))
                .andExpect(MockMvcResultMatchers.status().isOk())
                .andExpect(MockMvcResultMatchers.content().contentType(MediaType.APPLICATION_JSON));

    }

    @Test
    void addCustomer() {
    }

    @Test
    void getAll() {
    }

    @Test
    void deleteCustomerById() {
    }
}